babies/requirements.txt
Django==3.0.4
ipython==7.13.0
djangorestframework==3.11.0
